Track launchers using a dedicated type
This is quite similar to pkg/logs/scheduler.Schedulers.  One substantial
change is that the interesting internal bits of the logs-agent to which
a launcher has access are passed to its Start method.

This exposes the remaining "weird" launchers (container, kubernetes, and
docker) as feeling even weirder, but that will be addressed in
subsequent commits.