Replace the container launcher with a dedicated package
The current logic for deciding which launcher to run is complex, but
changing it will result in undesirable behavior changes.  This commit
isolates that logic in a dedicated containerorpods package and uses it
in the two launchers.

This prepares for breaking those launchers up into smaller components,
all of which will depend on the new package.